+++ src/main/java/com/nanwulife/service/IScoreService.java
@@ -12,7 +12,7 @@ public interface IScoreService {
 
     ServerResponse deleteScore(Integer stuId, Integer expId);
 
-    ServerResponse getScoreListByStunum(Integer userId, Integer expId, Integer isExport);
+    ServerResponse getScoreListByStunum(Long userId, Integer expId, Integer isExport);
 
     ServerResponse getScoreListByMajor(Integer majorId, Integer stuClass, Integer expId, Integer isExport, String orderBy);
